First stab at adding TEM to Field-to-Model Docker image.
Not sure if this is the approach that will be the best, but this was
able to work... I was able to build the image; building the image
compiles the dvmdostem executable. Then am able to launch container and
run the executable with the --help flag.

Need to more robust testing to ensure that all the aother libraries
linked properly...but this is a start.